src/migrations/versions/cd6cf66192b5_create_vehicle_table.py
"""create vehicle table

Revision ID: cd6cf66192b5
Revises: 
Create Date: 2023-10-08 20:31:27.595316

"""
from typing import Sequence, Union

from alembic import op
import sqlalchemy as sa


# revision identifiers, used by Alembic.
revision: str = 'cd6cf66192b5'
down_revision: Union[str, None] = None
branch_labels: Union[str, Sequence[str], None] = None
depends_on: Union[str, Sequence[str], None] = None


def upgrade() -> None:
    pass


def downgrade() -> None:
    pass
